dp9ik auth and cleanup
Pending merge upstream I am just going to bite the bullet and maintain
a personal branch for the time being. In the long term I will need to
do this anyway as divergences between unstable and nixos-23.05 have
nowhere else to go.

One thing considered was patching the upstream nixpkgs with a series
of patch files. This ended up being possible but caused what I learned
to be IFD. Which I conceptulize as forcing the entire nixpkgs repo
to evaluate. Which in practice would slow down flake builds. There
is a promising upstream feature in the works that would make this more
possible, but is currently a bit of a ways out it seems.